[IMP][O1758] separating hello world and authorisation steps
We need to be able to inherit the hello world module for other mtd
modules and use the authorisation steps.
