+++ account_mtd/models/__init__.py
@@ -21,12 +21,17 @@
 ##############################################################################
 
 from . import (
-    
-    mtd_hmrc_configuration,
+
     mtd_api_request_tracker,
     mtd_api_tokens,
     mtd_api,
+    mtd_construct_display_message,
+    mtd_exchange_authorisation,
     mtd_hello_world,
+    mtd_hmrc_configuration,
+    mtd_issue_request,
+    mtd_refresh_authorisation,
+    mtd_user_authorisation,
     
 )
