+++ account_mtd/models/mtd_api_request_tracker.py
@@ -20,7 +20,10 @@ class MtdApiRequestTracker(models.Model):
     closed = fields.Selection([
         ('timed_out', 'Request timed out'),
         ('response', 'Response received'),
-        ('more_than_one', 'More than one open tracker found')
+        ('more_than_one', 'More than one open tracker found'),
+        ('error', 'There was an error')
+
     ], string='Tracker Closed')
     action = fields.Char(required=True)
     menu_id = fields.Char(required=True)
+    module_name = fields.Char(required=True)
